+++ src/api/index.ts
@@ -1,6 +1,6 @@
 import combineRouters from 'koa-combine-routers'
-import postRouter from './post'
-import authRouter from './auth'
+import postRouter from './posts/post'
+import authRouter from './auth/auth'
 
 const router = combineRouters(postRouter, authRouter)
